Anima on frame revisions
Anima instinct remains unimplemented, though is now designated as something to ensure coherence.
Velocity set to a uint8, down from a uint32.
Syncing to the abstract has been moved to a (implementation local) fn.

+++ include/anima/abstract.h
@@ -10,6 +10,6 @@ struct abstract_anima_t {
 
   _Atomic(AnimaStatus) status;
 
-  _Atomic(uint32_t) speed;
+  _Atomic(uint8_t) speed;
 };
 typedef struct abstract_anima_t AbstractAnima;
